+++ ql/arch/cc_light_eqasm_compiler.h
@@ -9,6 +9,7 @@
 #ifndef QL_CC_LIGHT_EQASM_COMPILER_H
 #define QL_CC_LIGHT_EQASM_COMPILER_H
 
+#include <ql/utils.h>
 #include <ql/platform.h>
 #include <ql/kernel.h>
 #include <ql/gate.h>
@@ -151,6 +152,9 @@ public:
 
     size_t getRegNo( qubit_set_t & qs )
     {
+        // sort qubit operands to avoid variation in order
+        sort(qs.begin(), qs.end());
+
         auto it = QS2Mask.find(qs);
         if( it == QS2Mask.end() )
         {
@@ -163,6 +167,9 @@ public:
 
     size_t getRegNo( qubit_pair_set_t & qps )
     {
+        // sort qubit operands pair to avoid variation in order
+        sort(qps.begin(), qps.end(), ql::utils::sort_pair_helper);
+
         auto it = QPS2Mask.find(qps);
         if( it == QPS2Mask.end() )
         {
@@ -175,6 +182,9 @@ public:
 
     std::string getRegName( qubit_set_t & qs )
     {
+        // sort qubit operands to avoid variation in order
+        sort(qs.begin(), qs.end());
+
         auto it = QS2Mask.find(qs);
         if( it == QS2Mask.end() )
         {
@@ -187,6 +197,9 @@ public:
 
     std::string getRegName( qubit_pair_set_t & qps )
     {
+        // sort qubit operands pair to avoid variation in order
+        sort(qps.begin(), qps.end(), ql::utils::sort_pair_helper);
+
         auto it = QPS2Mask.find(qps);
         if( it == QPS2Mask.end() )
         {
@@ -390,13 +403,38 @@ std::string classical_instruction2qisa(ql::arch::classical_cc* classical_ins)
     return ssclassical.str();
 }
 
+
 std::string bundles2qisa(ql::ir::bundles_t & bundles,
-    ql::quantum_platform & platform, MaskManager & gMaskManager)
+    const ql::quantum_platform & platform, MaskManager & gMaskManager)
 {
     IOUT("Generating CC-Light QISA");
 
     std::stringstream ssbundles, sspre, ssinst;
     size_t curr_cycle=0;
+
+    // sort sections to get consistent output across multiple runs. The output
+    // is correct even without this sorting. Sorting is important to test the similarity
+    // of generated qisa against golden qisa files. For example, without sorting
+    // any of the following can be generated, which is correct but there will be
+    // differences reported by file_compare used for testing:
+    // x s0 | y s1
+    // OR
+    // y s1 | x s0
+    // However, with sorting it will always generate:
+    // x s0 | y s1
+    for (ql::ir::bundle_t & abundle : bundles)
+    {
+        // sorts instructions alphabetically
+        abundle.parallel_sections.sort( [] 
+            (const ql::ir::section_t & sec1, const ql::ir::section_t & sec2) -> bool
+            {
+                auto i1 = sec1.begin();
+                auto iname1 = (*(i1))->name;
+                auto i2 = sec2.begin();
+                auto iname2 = (*(i2))->name;
+                return iname2 < iname1;
+            });
+    }
 
     for (ql::ir::bundle_t & abundle : bundles)
     {
@@ -411,7 +449,8 @@ std::string bundles2qisa(ql::ir::bundles_t & bundles,
             sspre << "    qwait " << delta-1 << "\n"
                   << "    1    ";
 
-        for( auto secIt = abundle.parallel_sections.begin(); secIt != abundle.parallel_sections.end(); ++secIt )
+        for(auto secIt = abundle.parallel_sections.begin(); 
+            secIt != abundle.parallel_sections.end(); ++secIt )
         {
             qubit_set_t squbits;
             qubit_pair_set_t dqubits;
@@ -427,7 +466,7 @@ std::string bundles2qisa(ql::ir::bundles_t & bundles,
             else
             {
                 auto id = iname;
-                DOUT("get cclight instr name for : " << id);
+                // DOUT("get cclight instr name for : " << id);
                 std::string cc_light_instr_name;
                 auto it = platform.instruction_map.find(id);
                 if (it != platform.instruction_map.end())
@@ -471,12 +510,13 @@ std::string bundles2qisa(ql::ir::bundles_t & bundles,
                             throw ql::exception("Error : only 1 and 2 operand instructions are supported by cc light masks !",false);
                         }
                     }
+
                     std::string rname;
-                    if( 1 == nOperands )
+                    if(1 == nOperands)
                     {
                         rname = gMaskManager.getRegName(squbits);
                     }
-                    else if( 2 == nOperands )
+                    else if(2 == nOperands)
                     {
                         rname = gMaskManager.getRegName(dqubits);
                     }
@@ -527,7 +567,7 @@ std::string bundles2qisa(ql::ir::bundles_t & bundles,
 
     auto & lastBundle = bundles.back();
     int lbduration = lastBundle.duration_in_cycles;
-    if( lbduration>1 )
+    if(lbduration>1)
         ssbundles << "    qwait " << lbduration << "\n";
 
     IOUT("Generating CC-Light QISA [Done]");
@@ -594,7 +634,7 @@ void WriteCCLightQisaTimeStamped(std::string prog_name, ql::quantum_platform & p
             ssbundles << std::setw(8) << curr_cycle << ":    qwait " << delta-1 << "\n"
                       << std::setw(8) << curr_cycle + (delta-1) << ":    bs 1    ";
 
-        for( auto secIt = abundle.parallel_sections.begin(); secIt != abundle.parallel_sections.end(); ++secIt )
+        for(auto secIt = abundle.parallel_sections.begin(); secIt != abundle.parallel_sections.end(); ++secIt)
         {
             qubit_set_t squbits;
             qubit_pair_set_t dqubits;
@@ -604,7 +644,7 @@ void WriteCCLightQisaTimeStamped(std::string prog_name, ql::quantum_platform & p
             std::string cc_light_instr_name = get_cc_light_instruction_name(id, platform);
             auto itype = (*(firstInsIt))->type();
             auto nOperands = ((*firstInsIt)->operands).size();
-            if( itype == __nop_gate__ )
+            if(itype == __nop_gate__)
             {
                 ssbundles << cc_light_instr_name;
             }
@@ -645,7 +685,7 @@ void WriteCCLightQisaTimeStamped(std::string prog_name, ql::quantum_platform & p
                 ssbundles << cc_light_instr_name << " " << rname;
             }
 
-            if( std::next(secIt) != abundle.parallel_sections.end() )
+            if(std::next(secIt) != abundle.parallel_sections.end())
             {
                 ssbundles << " | ";
             }
@@ -691,7 +731,7 @@ public:
 
 
     /*
-     * program-level compilaation of qasm to cc_light_eqasm
+     * program-level compilation of qasm to cc_light_eqasm
      */
     void compile(std::string prog_name, ql::circuit& ckt, ql::quantum_platform& platform)
     {
@@ -729,8 +769,6 @@ public:
 
         // write scheduled bundles with parallelism in cc-light syntax with time-stamps
         WriteCCLightQisaTimeStamped(prog_name, platform, mask_manager, bundles);
-
-
 
 
         // time analysis
@@ -849,11 +887,27 @@ public:
     }
 
     // kernel level compilation
-    void compile(std::string prog_name, std::vector<quantum_kernel> kernels, ql::quantum_platform& platform)
+    void compile(std::string prog_name, std::vector<quantum_kernel> kernels, const ql::quantum_platform& platform)
     {
         DOUT("Compiling " << kernels.size() << " kernels to generate CCLight eQASM ... ");
 
         load_hw_settings(platform);
+
+        const json& instruction_settings = platform.instruction_settings;
+        for(const json & i : instruction_settings)
+        {
+            std::string instr_name;
+            if(i.count("cc_light_instr") <= 0)
+            {
+                EOUT("cc_light_instr not found for " << i);
+                throw ql::exception("cc_light_instr not found", false);
+            }
+            else
+            {
+                instr_name = i["cc_light_instr"];
+            }
+        }
+
         generate_opcode_cs_files(platform);
         MaskManager mask_manager;
 
@@ -911,7 +965,7 @@ public:
     /**
      * decompose
      */
-    void decompose_instructions(ql::circuit& ckt, ql::circuit& decomp_ckt, ql::quantum_platform& platform)
+    void decompose_instructions(ql::circuit& ckt, ql::circuit& decomp_ckt, const ql::quantum_platform& platform)
     {
         DOUT("decomposing instructions...");
         for( auto ins : ckt )
@@ -968,7 +1022,7 @@ public:
                 }
                 else
                 {
-                    json& instruction_settings = platform.instruction_settings;
+                    const json& instruction_settings = platform.instruction_settings;
                     std::string operation_type;
                     if (instruction_settings.find(iname) != instruction_settings.end())
                     {
@@ -1211,7 +1265,7 @@ public:
 
 private:
 
-    void load_hw_settings(ql::quantum_platform& platform)
+    void load_hw_settings(const ql::quantum_platform& platform)
     {
         std::string params[] = { "qubit_number", "cycle_time", "mw_mw_buffer", "mw_flux_buffer", "mw_readout_buffer", "flux_mw_buffer",
                                  "flux_flux_buffer", "flux_readout_buffer", "readout_mw_buffer", "readout_flux_buffer", "readout_readout_buffer"
@@ -1240,10 +1294,10 @@ private:
         }
     }
 
-    void generate_opcode_cs_files(ql::quantum_platform& platform)
+    void generate_opcode_cs_files(const ql::quantum_platform& platform)
     {
         DOUT("Generating opcode file ...");
-        json& instruction_settings       = platform.instruction_settings;
+        const json& instruction_settings       = platform.instruction_settings;
 
         std::stringstream opcode_ss;
 
@@ -1281,24 +1335,24 @@ private:
         std::map<std::string,size_t> instr_name_2_opcode;
         std::set<size_t> opcode_set;
         size_t opcode=0;
-        for (json & i : instruction_settings)
+        for (const json & i : instruction_settings)
         {
             std::string instr_name;
-            if (i["cc_light_instr"].is_null())
+            // COUT("Looking for instruction: " << i);
+            if (i.count("cc_light_instr") <= 0)
             {
                 EOUT("cc_light_instr not found for " << i);
-                throw ql::exception("cc_light_instr not found", false);
+                throw ql::exception("cc_light_instr not found for <> ", false);
             }
             else
             {
                 instr_name = i["cc_light_instr"];
             }
 
-            if (i["cc_light_opcode"].is_null())
+            if (i.count("cc_light_opcode") <= 0)
                 throw ql::exception("[x] error : ql::eqasm_compiler::compile() : missing opcode for instruction '"+instr_name,false);
             else
                 opcode = i["cc_light_opcode"];
-
 
             auto mapit = instr_name_2_opcode.find(instr_name);
             if( mapit != instr_name_2_opcode.end() )
@@ -1329,9 +1383,12 @@ private:
                     throw ql::exception("[x] error : ql::eqasm_compiler::compile() : invalid opcode for single qubit gate instruction '"+instr_name+"' : should be in [1..127] range : current opcode: "+std::to_string(opcode),false);
                 }
                 opcode_set.insert(opcode);
-                size_t condition  = (i["cc_light_cond"].is_null() ? 0 : i["cc_light_cond"].get<size_t>());
-                if (i["cc_light_instr"].is_null())
+
+                size_t condition  = (i.count("cc_light_cond")<=0? 0 :i["cc_light_cond"].get<size_t>());
+
+                if (i.count("cc_light_instr") <=0 )
                     throw ql::exception("[x] error : ql::eqasm_compiler::compile() : 'cc_light_instr' attribute missing in gate definition (opcode: "+std::to_string(opcode),false);
+
                 opcode_ss << "def_q_arg_st[" << i["cc_light_instr"] << "]\t= " << std::showbase << std::hex << opcode << "\n";
                 auto optype     = (i["type"] == "mw" ? 1 : (i["type"] == "flux" ? 2 : ((i["type"] == "readout" ? 3 : 0))));
                 auto codeword   = i["cc_light_codeword"];
@@ -1345,9 +1402,10 @@ private:
                 if (opcode < 127 || opcode > 255)
                     throw ql::exception("[x] error : ql::eqasm_compiler::compile() : invalid opcode for two qubits gate instruction '"+instr_name+"' : should be in [128..255] range : current opcode: "+std::to_string(opcode),false);
                 opcode_set.insert(opcode);
-                // size_t condition  = 0;
-                size_t condition  = (i["cc_light_cond"].is_null() ? 0 : i["cc_light_cond"].get<size_t>());
-                if (i["cc_light_instr"].is_null())
+
+                size_t condition  = (i.count("cc_light_cond") <= 0? 0 :i["cc_light_cond"].get<size_t>());
+
+                if (i.count("cc_light_instr") <= 0)
                     throw ql::exception("[x] error : ql::eqasm_compiler::compile() : 'cc_light_instr' attribute missing in gate definition (opcode: "+std::to_string(opcode),false);
                 // opcode_ss << "def_opcode[" << i["cc_light_instr"] << "]\t= " << opcode << "\n";
                 opcode_ss << "def_q_arg_tt[" << i["cc_light_instr"] << "]\t= " << std::showbase << std::hex << opcode << "\n";
@@ -1358,8 +1416,6 @@ private:
             }
             else
                 throw ql::exception("[x] error : ql::eqasm_compiler::compile() : error while reading hardware settings : invalid 'cc_light_instr_type' for instruction !",false);
-            // println("\n" << control_store.str());
-            // println("\n" << opcode_ss.str());
         }
 
         std::string cs_filename = ql::options::get("output_dir") + "/cs.txt";