+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [ 0.5.5 ] - [ 2018-10-25 ]
+### Added
+
+### Changed
+- simplified interface of Program.set_sweep_points (Issue #184)
+### Removed
+
+### Fixed
+- instruction ordering to generate consistent qisa (Issue #190)
+- stateful behaviour in OpenQL (Issue #171)
+
+
+## [ 0.5.4 ] - [ 2018-10-17 ]
+### Added
+
+### Changed
+
+### Removed
+
+### Fixed
+- qubit ordering in SMIS and SMIT instructions
+
+
 ## [ 0.5.3 ] - [ 2018-10-11 ]
 ### Added
 - added detuning constraints for cclight